pyproject.toml
[tool.poetry]
name = "django-react-test"
version = "0.1.0"
description = ""
authors = ["nestor <albquenes7294@gmail.com>"]
readme = "README.md"
packages = [{include = "django_react_test"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
